+++ README.md
@@ -12,6 +12,7 @@ This package is used by [Converter NOW](https://github.com/ferraridamiano/Conver
 The documentation is structured in examples of incresing complexity. But don't worry, in most cases you will only need the first examples.
 
 ## Table of Contents
+- [Convert a unit to another unit (the easy way)](#convert-a-unit-to-another-unit-the-easy-way)
 - [Convert a unit to another unit](#convert-a-unit-to-another-unit)
 - [Convert a unit to all the another units](#convert-a-unit-to-all-the-another-units)
 - [Special type of conversion: numeral systems conversion](#special-type-of-conversion-numeral-systems-conversion)
@@ -19,8 +20,41 @@ The documentation is structured in examples of incresing complexity. But don't w
 - [Custom conversion](#custom-conversion)
 - [Which conversions?](#which-conversions)
 
-## Convert a unit to another unit
+## Convert a unit to another unit (the easy way)
 **Example 1**: convert 1 meter in inches
+
+```dart
+// Convert 1 meter in inches
+var inches = 1.convertFromTo(LENGTH.meters, LENGTH.inches);
+// Print it
+print(inches);
+```
+
+Output:
+
+```
+39.37007874015748
+```
+
+What about numeral systems? Same thing but with `String`:
+
+```dart
+// Convert '8B' hexadecimal to octal
+var octal = '8B'.convertFromTo(NUMERAL_SYSTEMS.hexadecimal, NUMERAL_SYSTEMS.octal);
+// Print it
+print(octal);
+```
+
+Output:
+
+```
+213
+```
+
+**Note**: most of the time is what you need, use this method if you just want to convert a unit in another unit. BUT, it is not recommended if you want to convert a unit to multiple units, or if you are interested in the symbol of a unit or if you are interested in the format of the string representation of the value (significant figures, scientific notation, etc.).
+
+## Convert a unit to another unit
+**Example 2**: convert 1 meter in inches
 
 ```dart
 // We give 1 meter as input
@@ -38,7 +72,7 @@ name:LENGTH.inches, value:39.370078740157474, stringValue:39.37007874, symbol:in
 ```
 
 ## Convert a unit to all the another units
-**Example 2**: convert 1 degree in all the other angles units. This time we want also to specify that we just want 7 significant figures and we don't want trailing zeros (e.g. 1.000000 -> 1).
+**Example 3**: convert 1 degree in all the other angles units. This time we want also to specify that we just want 7 significant figures and we don't want trailing zeros (e.g. 1.000000 -> 1).
 
 ```dart
 // Initialization of the object
@@ -65,7 +99,7 @@ As you can see in this example if you specify `removeTrailingZeros: false`, the 
 ## Special type of conversion: numeral systems conversion
 *Warning! Numeral systems conversion is the only conversion that need the input as a `String`, and not as a `double` / `int` for obvious reasons*
 
-**Example 3**: convert '`100`' (decimal) to binary and hexadecimal
+**Example 4**: convert '`100`' (decimal) to binary and hexadecimal
 
 ```dart
 // We give '100' decimal as input
@@ -86,7 +120,7 @@ Hexadecimal: 64
 ## Simple custom conversion
 *Use `SimpleCustomConversion` when you are dealing with a linear conversion between the units, i.e. when a unit is `x` times a value. In `SimpleCustomConversion` we have to define a conversionMap between a base unit, which must have a value of 1, and all the other units. In the example below we say that 1€ is 1.2271$, but also 0.9033₤, and so on and so forth.*
 
-**Example 4**: define the currency exchange rate with their symbols (optional) and perform the conversion between two of them.
+**Example 5**: define the currency exchange rate with their symbols (optional) and perform the conversion between two of them.
 
 ```dart
 final Map<String, double> conversionMap = {
@@ -121,7 +155,7 @@ Output:
 
 *In most cases, you will only need `SimpleCustomConversion` (see the previous section). `SimpleCustomConversion` allow you to define conversions in the form of `y=ax`. But if you need to define special relationship between units you need `CustomConversion`. This allow you to perform conversion like: `y=ax+b` and `y=a/x+b` (where `y` and x are the value of two units and `a` and `b` are two coefficient), for example the conversion between Celsius and Fahreneit use the first relation and the conversion between km/l and l/100km has to be done with the second relation. Both can't be done with `SimpleCustomConversion`.*
 
-**Example 5**: let's define the following imaginary units of measurement:
+**Example 6**: let's define the following imaginary units of measurement:
 
 * `KiloDash` = 1000 * `Dash`
 * `DashPlus1` = `Dash` + 1
