+++ example/example3.dart
@@ -1,11 +1,15 @@
 import 'package:units_converter/units_converter.dart';
 
 void main() {
-  // We give '100' decimal as input
-  var numeralSystems = NumeralSystems()
-    ..convert(NUMERAL_SYSTEMS.decimal, '100');
-  // We get the binary value
-  print('Binary: ${numeralSystems.binary.stringValue}');
-  // We get the hexadecimal value
-  print('Hexadecimal: ${numeralSystems.hexadecimal.stringValue}');
+  // Initialization of the object
+  var angle =
+      Angle(significantFigures: 7, removeTrailingZeros: false); // conversion
+  angle.convert(ANGLE.degree, 1);
+  // We get all the units
+  var units = angle.getAll(); //We get all ther others units
+  // Let's print all of them
+  for (var unit in units) {
+    print(
+        'name:${unit.name}, value:${unit.value}, stringValue:${unit.stringValue}, symbol:${unit.symbol}');
+  }
 }
